+++ packages/apps/crm/src/screens/prospect/ProspectDetailsScreen.js
@@ -1,55 +1,21 @@
-import React, {useEffect, useMemo, useCallback} from 'react';
+import React, {useEffect} from 'react';
 import {ScrollView, StyleSheet, View} from 'react-native';
+import {Screen, HeaderContainer, CircleButton} from '@axelor/aos-mobile-ui';
 import {
-  Screen,
-  HeaderContainer,
-  useThemeColor,
-  Text,
-  Badge,
-  StarScore,
-  DropdownCardSwitch,
-  NotesCard,
-  LabelText,
-  CircleButton,
-} from '@axelor/aos-mobile-ui';
-import {
-  useTranslator,
   useSelector,
   HeaderOptionsMenu,
   useDispatch,
-  AOSImageBubble,
 } from '@axelor/aos-mobile-core';
-import {
-  DropdownContactView,
-  DropdownEventView,
-  LiteContactCard,
-  ProspectHeader,
-} from '../../components';
+import {ProspectBody, ProspectHeader} from '../../components';
 import {searchContactById} from '../../features/contactSlice';
 import {fetchPartnerEventById} from '../../features/eventSlice';
-import {getLastEvent, getNextEvent} from '../../utils/dateEvent';
-import {
-  fetchProspectById,
-  updateProspectScore,
-} from '../../features/prospectSlice';
+import {fetchProspectById} from '../../features/prospectSlice';
 
 const ProspectDetailsScreen = ({navigation, route}) => {
   const idProspect = route.params.idProspect;
-  const I18n = useTranslator();
-  const Colors = useThemeColor();
   const dispatch = useDispatch();
   const {mobileSettings} = useSelector(state => state.config);
-  const {listContactById} = useSelector(state => state.contact);
-  const {listEventPartner} = useSelector(state => state.event);
   const {prospect} = useSelector(state => state.prospect);
-
-  const lastEventProspect = useMemo(() => {
-    return getLastEvent(listEventPartner);
-  }, [listEventPartner]);
-
-  const nextEventProspect = useMemo(() => {
-    return getNextEvent(listEventPartner);
-  }, [listEventPartner]);
 
   React.useLayoutEffect(() => {
     navigation.setOptions({
@@ -78,19 +44,6 @@ const ProspectDetailsScreen = ({navigation, route}) => {
     dispatch(fetchPartnerEventById(idProspect));
   }, [dispatch, idProspect]);
 
-  const updateScoreProspectAPI = useCallback(
-    newScore => {
-      dispatch(
-        updateProspectScore({
-          partnerId: prospect.id,
-          partnerVersion: prospect.version,
-          newScore: newScore,
-        }),
-      );
-    },
-    [dispatch, prospect.id, prospect.version],
-  );
-
   return (
     <Screen removeSpaceOnTop={true}>
       <HeaderContainer
@@ -98,96 +51,7 @@ const ProspectDetailsScreen = ({navigation, route}) => {
         fixedItems={<ProspectHeader />}
       />
       <ScrollView>
-        <NotesCard title={I18n.t('Crm_Notes')} data={prospect.description} />
-        <View style={styles.container}>
-          <DropdownCardSwitch
-            styleTitle={styles.textTitle}
-            dropdownItems={[
-              {
-                title: I18n.t('Crm_Contact'),
-                key: 1,
-                childrenComp: (
-                  <DropdownContactView
-                    address={prospect.mainAddress?.fullName}
-                    fixedPhone={prospect.fixedPhone}
-                    emailAddress={prospect.emailAddress?.address}
-                    webSite={prospect.webSite}
-                  />
-                ),
-              },
-              {
-                title: I18n.t('Crm_GeneralInformation'),
-                key: 2,
-                childrenComp: (
-                  <View>
-                    {prospect.user?.fullName && (
-                      <LabelText
-                        title={I18n.t('Crm_AssignedTo')}
-                        iconName={'user-tie'}
-                        value={prospect.user?.fullName}
-                      />
-                    )}
-                    {prospect.type?.name && (
-                      <LabelText
-                        title={I18n.t('Crm_Categorie')}
-                        value={prospect.partnerCategory?.name}
-                      />
-                    )}
-                    {prospect.industrySector?.name && (
-                      <LabelText
-                        title={I18n.t('Crm_Sector')}
-                        value={prospect.industrySector?.name}
-                      />
-                    )}
-                    {!prospect.user?.fullName &&
-                      !prospect.type?.name &&
-                      !prospect.industrySector?.name && (
-                        <View>
-                          <Text>{I18n.t('Crm_NoGeneralInformation')}</Text>
-                        </View>
-                      )}
-                  </View>
-                ),
-              },
-              {
-                title: I18n.t('Crm_Employees'),
-                key: 3,
-                childrenComp:
-                  listContactById.length > 0 ? (
-                    listContactById.map((contact, index) => (
-                      <LiteContactCard
-                        key={index}
-                        contactFullname={contact.simpleFullName}
-                        fixedPhoneNumber={contact.fixedPhone}
-                        email={contact['emailAddress.address']}
-                        style={styles.item}
-                        onPress={() =>
-                          navigation.navigate('ContactDetailsScreen', {
-                            idContact: contact.id,
-                          })
-                        }
-                      />
-                    ))
-                  ) : (
-                    <View>
-                      <Text>{I18n.t('Crm_NoContactAssociated')}</Text>
-                    </View>
-                  ),
-              },
-              {
-                title: I18n.t('Crm_Events'),
-                key: 4,
-                childrenComp: (
-                  <DropdownEventView
-                    lastEvent={lastEventProspect}
-                    nextEvent={nextEventProspect}
-                    navigation={navigation}
-                  />
-                ),
-              },
-            ]}
-          />
-        </View>
+        <ProspectBody navigation={navigation} />
       </ScrollView>
       <View style={styles.bottomContainer}>
         <CircleButton
